tests: check for regression of overflow fix
Commit a6d9848a32fafa763548e54b44cb094abdac915d fixed an
integer overflow when parsing a numerical value given to
the send do, send dont, send will, and send wont commands.
Check that overly large or small values are caught and
reported as 'bad value' even if they cannot be stored in
a 64-bit integer.

* tests/telnet-localhost.sh: New check to detect regressions
in the recent integer overflow fix.