maint: Use AC_CONFIG_MACRO_DIRS instead of ACLOCAL_AMFLAGS.
Also arrange order of directories so that am__aclocal_m4_deps
in Makefile.in is generated the same by ./bootstrap and
autoreconf.  The latter is invoked by 'make release' (via the
_version rule in GNUmakefile) to change the version number.
We want am__aclocal_m4_deps to be identical so that the
"make dist" and "make release" tarballs are identical.

* Makefile.am (ACLOCAL_AMFLAGS): Remove.
configure.ac (AC_CONFIG_MACRO_DIRS): Add.
bootstrap.conf (ACLOCAL_AMFLAGS): Remove.