tests: Remove bogus libls test for unsorted file listing
We cannot reliably test whether the -f option works against a normal
filesystem, because that relies on the unsorted output coming out
accidentally not sorted, and this has been the cause for several
indeterministic build failures in various hosts (such as some
sparc64 or reproducible build nodes).

This could be guaranteed with something like disorderfs, but we do
not bother and simply remove the test case.

Signed-off-by: Simon Josefsson <simon@josefsson.org>