syslod, talkd: GNU/Linux w/o utmp compatibility
Some newer GNU/Linux systems no longer have a "utmp" file.
This file is used by both syslogd and talkd to find the
TTY device files of active user sessions in order to send
messages to specific or all users.

Gnulib can emulate a utmp file via the read_utmp() function
from the readutmp module.  This emulation is activated by
defining "READUTMP_USE_SYSTEMD", triggered via configure's
"--enable-systemd" option on systems without utmp and with
systemd.

But, this emulation reports non-existing TTY device files
for active user sessions.  Attempting to send a message to a
non-existing TTY produces an error, but that is normal behavior
for such a system.  Thus, ignore this error on such systems,
but not others.

* NEWS.md: Mention enhancement.
* libinetutils/ttymsg.c (inetutils_ttymsg): Ignore errors for
  non-existings TTY files reported in emulated utmp entries.